--- src/main/java/org/programmer/kdtspringboot/io/Console.java
@@ -1,26 +1,0 @@
-package org.programmer.kdtspringboot.io;
-
-import org.programmer.kdtspringboot.io.Input;
-import org.programmer.kdtspringboot.io.Output;
-
-public class Console implements Input, Output {
-    @Override
-    public String input(String s) {
-        return null;
-    }
-
-    @Override
-    public void menu() {
-
-    }
-
-    @Override
-    public void print(String content) {
-
-    }
-
-    @Override
-    public void exit() {
-
-    }
-}
